Reordering modified so eliminate repetition per semester

+++ SPOT/Actions/ActionShowCourseInfo.cpp
@@ -75,6 +75,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 1 || pC->getsemester() != FALL)
 					{
+						if (courseAlreadyExists(pC, 1, FALL))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 1, FALL);
 						pC->setyear(1);
@@ -86,6 +92,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 1 || pC->getsemester() != SPRING)
 					{
+						if (courseAlreadyExists(pC, 1, SPRING))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 1, SPRING);
 						pC->setyear(1);
@@ -98,6 +110,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 1 || pC->getsemester() != SUMMER)
 					{
+						if (courseAlreadyExists(pC, 1, SUMMER))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pC->setyear(1);
 						pC->setsemester(SUMMER);
@@ -110,6 +128,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 2 || pC->getsemester() != FALL)
 					{
+						if (courseAlreadyExists(pC, 2, FALL))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 2, FALL);
 						pC->setyear(2);
@@ -121,6 +145,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 2 || pC->getsemester() != SPRING)
 					{
+						if (courseAlreadyExists(pC, 2, SPRING))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 2, SPRING);
 						pC->setyear(2);
@@ -132,6 +162,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 2 || pC->getsemester() != SUMMER)
 					{
+						if (courseAlreadyExists(pC, 2, SUMMER))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 2, SUMMER);
 						pC->setyear(2);
@@ -143,6 +179,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 3 || pC->getsemester() != FALL)
 					{
+						if (courseAlreadyExists(pC, 3, FALL))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 3, FALL);
 						pC->setyear(3);
@@ -154,6 +196,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 3 || pC->getsemester() != SPRING)
 					{
+						if (courseAlreadyExists(pC, 3, SPRING))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 3, SPRING);
 						pC->setyear(3);
@@ -165,6 +213,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 3 || pC->getsemester() != SUMMER)
 					{
+						if (courseAlreadyExists(pC, 3, SUMMER))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 3, SUMMER);
 						pC->setyear(3);
@@ -176,6 +230,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 4 || pC->getsemester() != FALL)
 					{
+						if (courseAlreadyExists(pC, 4, FALL))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 4, FALL);
 						pC->setyear(4);
@@ -187,6 +247,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 4 || pC->getsemester() != SPRING)
 					{
+						if (courseAlreadyExists(pC, 4, SPRING))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 4, SPRING);
 						pC->setyear(4);
@@ -198,6 +264,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 4 || pC->getsemester() != SUMMER)
 					{
+						if (courseAlreadyExists(pC, 4, SUMMER))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 4, SUMMER);
 						pC->setyear(4);
@@ -209,6 +281,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 5 || pC->getsemester() != FALL)
 					{
+						if (courseAlreadyExists(pC, 5, FALL))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 5, FALL);
 						pC->setyear(5);
@@ -220,6 +298,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 5 || pC->getsemester() != SPRING)
 					{
+						if (courseAlreadyExists(pC, 5, SPRING))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 5, SPRING);
 						pC->setyear(5);
@@ -231,6 +315,12 @@ bool ActionShowCourseInfo::Execute()
 				{
 					if (pC->getyear() != 5 || pC->getsemester() != SUMMER)
 					{
+						if (courseAlreadyExists(pC, 5, SUMMER))
+						{
+							pC->setSelected(false);
+							return false;
+						}
+
 						pS->DeleteCourse(pC);
 						pS->AddCourse(pC, 5, SUMMER);
 						pC->setyear(5);
@@ -349,8 +439,20 @@ void ActionShowCourseInfo::showInfo(window* windptr, Course* pC)
 
 bool ActionShowCourseInfo::courseAlreadyExists(Course* pC, int year, SEMESTER sem)
 {
-	
-	return false;
+	//returns true if the course is already there and false if the course is not there
+	Course* pCR = pReg->getStudyPlay()->searchSemester(pC->getCode(), year, sem);
+	if (pCR) {
+		window* pW = pReg->getGUI()->getPwind();
+		pCR->setSelected(true);
+		pReg->UpdateInterface();
+		ActionShowCourseInfo::showInfo(pW, pC);
+		pReg->getGUI()->GetUserAction("Course already added in this semester. press anywhere to dismiss");
+
+		pCR->setSelected(false);
+		return true;
+	}
+	else
+		return false;
 }
 
 ActionShowCourseInfo::~ActionShowCourseInfo()
